Separate setValue to setValue and setValueByPath.
