+++ src/ArrayHelper.php
@@ -301,6 +301,64 @@ class ArrayHelper
      *  ];
      * ```
      *
+     * The result of `ArrayHelper::setValue($array, ['key', 'in'], ['arr' => 'val']);`
+     * will be the following:
+     *
+     * ```php
+     *  [
+     *      'key' => [
+     *          'in' => [
+     *              'arr' => 'val'
+     *          ]
+     *      ]
+     *  ]
+     * ```
+     *
+     * @param array $array the array to write the value to
+     * @param string|array|null $key the path of where do you want to write a value to `$array`
+     * the path can be described by an array of keys
+     * if the path is null then `$array` will be assigned the `$value`
+     * @param mixed $value the value to be written
+     */
+    public static function setValue(array &$array, $key, $value): void
+    {
+        if ($key === null) {
+            $array = $value;
+            return;
+        }
+
+        $keys = is_array($key) ? $key : [$key];
+
+        while (count($keys) > 1) {
+            $k = array_shift($keys);
+            if (!isset($array[$k])) {
+                $array[$k] = [];
+            }
+            if (!is_array($array[$k])) {
+                $array[$k] = [$array[$k]];
+            }
+            $array = &$array[$k];
+        }
+
+        $array[array_shift($keys)] = $value;
+    }
+
+    /**
+     * Writes a value into an associative array at the key path specified.
+     * If there is no such key path yet, it will be created recursively.
+     * If the key exists, it will be overwritten.
+     *
+     * ```php
+     *  $array = [
+     *      'key' => [
+     *          'in' => [
+     *              'val1',
+     *              'key' => 'val'
+     *          ]
+     *      ]
+     *  ];
+     * ```
+     *
      * The result of `ArrayHelper::setValue($array, 'key.in.0', ['arr' => 'val']);` will be the following:
      *
      * ```php
@@ -337,27 +395,12 @@ class ArrayHelper
      * if the path is null then `$array` will be assigned the `$value`
      * @param mixed $value the value to be written
      */
-    public static function setValue(array &$array, $path, $value): void
+    public static function setValueByPath(array &$array, $path, $value): void
     {
-        if ($path === null) {
-            $array = $value;
-            return;
+        if (is_string($path)) {
+            $path = explode('.', $path);
         }
-
-        $keys = is_array($path) ? $path : explode('.', $path);
-
-        while (count($keys) > 1) {
-            $key = array_shift($keys);
-            if (!isset($array[$key])) {
-                $array[$key] = [];
-            }
-            if (!is_array($array[$key])) {
-                $array[$key] = [$array[$key]];
-            }
-            $array = &$array[$key];
-        }
-
-        $array[array_shift($keys)] = $value;
+        static::setValue($array, $path, $value);
     }
 
     /**
