+++ tests/ArrayHelperTest.php
@@ -632,261 +632,6 @@ final class ArrayHelperTest extends TestCase
         ArrayHelper::getValue($object, 'var');
     }
 
-    /**
-     * Data provider for [[testSetValue()]].
-     * @return array test data
-     */
-    public function dataProviderSetValue(): array
-    {
-        return [
-            [
-                [
-                    'key1' => 'val1',
-                    'key2' => 'val2',
-                ],
-                'key',
-                'val',
-                [
-                    'key1' => 'val1',
-                    'key2' => 'val2',
-                    'key' => 'val',
-                ],
-            ],
-            [
-                [
-                    'key1' => 'val1',
-                    'key2' => 'val2',
-                ],
-                'key2',
-                'val',
-                [
-                    'key1' => 'val1',
-                    'key2' => 'val',
-                ],
-            ],
-
-            [
-                [
-                    'key1' => 'val1',
-                ],
-                'key.in',
-                'val',
-                [
-                    'key1' => 'val1',
-                    'key' => ['in' => 'val'],
-                ],
-            ],
-            [
-                [
-                    'key' => 'val1',
-                ],
-                'key.in',
-                'val',
-                [
-                    'key' => [
-                        'val1',
-                        'in' => 'val',
-                    ],
-                ],
-            ],
-            [
-                [
-                    'key' => 'val1',
-                ],
-                'key',
-                ['in' => 'val'],
-                [
-                    'key' => ['in' => 'val'],
-                ],
-            ],
-
-            [
-                [
-                    'key1' => 'val1',
-                ],
-                'key.in.0',
-                'val',
-                [
-                    'key1' => 'val1',
-                    'key' => [
-                        'in' => ['val'],
-                    ],
-                ],
-            ],
-
-            [
-                [
-                    'key1' => 'val1',
-                ],
-                'key.in.arr',
-                'val',
-                [
-                    'key1' => 'val1',
-                    'key' => [
-                        'in' => [
-                            'arr' => 'val',
-                        ],
-                    ],
-                ],
-            ],
-            [
-                [
-                    'key1' => 'val1',
-                ],
-                'key.in.arr',
-                ['val'],
-                [
-                    'key1' => 'val1',
-                    'key' => [
-                        'in' => [
-                            'arr' => ['val'],
-                        ],
-                    ],
-                ],
-            ],
-            [
-                [
-                    'key' => [
-                        'in' => ['val1'],
-                    ],
-                ],
-                'key.in.arr',
-                'val',
-                [
-                    'key' => [
-                        'in' => [
-                            'val1',
-                            'arr' => 'val',
-                        ],
-                    ],
-                ],
-            ],
-
-            [
-                [
-                    'key' => ['in' => 'val1'],
-                ],
-                'key.in.arr',
-                ['val'],
-                [
-                    'key' => [
-                        'in' => [
-                            'val1',
-                            'arr' => ['val'],
-                        ],
-                    ],
-                ],
-            ],
-            [
-                [
-                    'key' => [
-                        'in' => [
-                            'val1',
-                            'key' => 'val',
-                        ],
-                    ],
-                ],
-                'key.in.0',
-                ['arr' => 'val'],
-                [
-                    'key' => [
-                        'in' => [
-                            ['arr' => 'val'],
-                            'key' => 'val',
-                        ],
-                    ],
-                ],
-            ],
-            [
-                [
-                    'key' => [
-                        'in' => [
-                            'val1',
-                            'key' => 'val',
-                        ],
-                    ],
-                ],
-                'key.in',
-                ['arr' => 'val'],
-                [
-                    'key' => [
-                        'in' => ['arr' => 'val'],
-                    ],
-                ],
-            ],
-            [
-                [
-                    'key' => [
-                        'in' => [
-                            'key' => 'val',
-                            'data' => [
-                                'attr1',
-                                'attr2',
-                                'attr3',
-                            ],
-                        ],
-                    ],
-                ],
-                'key.in.schema',
-                'array',
-                [
-                    'key' => [
-                        'in' => [
-                            'key' => 'val',
-                            'schema' => 'array',
-                            'data' => [
-                                'attr1',
-                                'attr2',
-                                'attr3',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-            [
-                [
-                    'key' => [
-                        'in.array' => [
-                            'key' => 'val',
-                        ],
-                    ],
-                ],
-                ['key', 'in.array', 'ok.schema'],
-                'array',
-                [
-                    'key' => [
-                        'in.array' => [
-                            'key' => 'val',
-                            'ok.schema' => 'array',
-                        ],
-                    ],
-                ],
-            ],
-            [
-                [
-                    'key' => ['val'],
-                ],
-                null,
-                'data',
-                'data',
-            ],
-        ];
-    }
-
-    /**
-     * @dataProvider dataProviderSetValue
-     *
-     * @param array $arrayInput
-     * @param string|array|null $key
-     * @param mixed $value
-     * @param mixed $expected
-     */
-    public function testSetValue(array $arrayInput, $key, $value, $expected): void
-    {
-        ArrayHelper::setValue($arrayInput, $key, $value);
-        $this->assertEquals($expected, $arrayInput);
-    }
-
     public function testIsAssociative(): void
     {
         $this->assertFalse(ArrayHelper::isAssociative([]));
